debug

+++ improved_diffusion/text_nn.py
@@ -407,8 +407,8 @@ class CrossAttention(nn.Module):
                    'tmn', f"{max_over_src.min().item():.4f}",
                    ))
 
-        tgt_norm1 = (tgt ** 2).sum().sqrt().item()
-        attn_norm = (attn_output ** 2).sum().sqrt().item()
+        tgt_norm1 = (tgt.float() ** 2).sum().sqrt().item()
+        attn_norm = (attn_output.float() ** 2).sum().sqrt().item()
 
         print(("tgt_norm1", tgt_norm1))
         print(("attn_norm", attn_norm))
